+++ scala-package/examples/src/test/scala/org/apache/mxnetexamples/customop/CustomOpExampleSuite.scala
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.mxnet.examples.customop
+package org.apache.mxnetexamples.customop
 
 import java.io.File
 import java.net.URL
